+++ Runtime/Exceptions/AffiseError.cs
@@ -3,8 +3,10 @@ using System;
 
 namespace AffiseAttributionLib.Exceptions
 {
-    internal abstract class AffiseError : Exception
+    public abstract class AffiseError : Exception
     {
-        public const string MESSAGE_ALREADY_INITIALIZED = "Affise SDK is already initialized";
+        internal const string MESSAGE_ALREADY_INITIALIZED = "Affise SDK is already initialized";
+        public const string UUID_NOT_INITIALIZED = "11111111-1111-1111-1111-111111111111";
+        public const string UUID_NO_VALID_METHOD = "22222222-2222-2222-2222-222222222222";
     }
 }